+++ NEWS.md
@@ -2,7 +2,7 @@
  * The order of dataset `ST` is now based on the full subgroup level code (https://github.com/ncss-tech/SoilTaxonomy/issues/35). Partial codes corresponding to each level of the hierarchy are also included in columns: `order_code`, `suborder_code`, `greatgroup_code` and `subgroup_code`.
  * New function `SoilTaxonomyLevels()`: used for creating ordinal and nominal factors corresponding to taxonomic levels `"order"`, `"suborder"`, `"greatgroup"`, and `"subgroup"` (https://github.com/ncss-tech/SoilTaxonomy/issues/39).
  * `FormativeElements()` can now take multiple `level` values, returning the formative elements found at each level combined in a single data.frame.
- 
+ * New function `get_ST_formative_elements()` a helper function for getting a data.frame containing descriptors of formative elements used at the specified `level`(s)
 
 # SoilTaxonomy 0.2.1 (2022-07-21)
  * New function `newick_string()`: used for creating parenthetical "Newick" (or "New Hampshire") format tree string inputs to functions such as `ape::read.tree()`